reorganize some code
