+++ app/actions/index.js
@@ -1,6 +1,6 @@
 import uuid from 'node-uuid'
 
-export const addTodo = (text) => {
+export const addTodo = text => {
   return {
     type: 'ADD_TODO',
     id: uuid.v4(),
@@ -8,14 +8,14 @@ export const addTodo = (text) => {
   }
 }
 
-export const setVisibilityFilter = (filter) => {
+export const setVisibilityFilter = filter => {
   return {
     type: 'SET_VISIBILITY_FILTER',
     filter
   }
 }
 
-export const toggleTodo = (id) => {
+export const toggleTodo = id => {
   return {
     type: 'TOGGLE_TODO',
     id
