+++ app/components/FilterLinks.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import FilterLink from './FilterLink'
+import FilterLink from '../containers/FilterLink'
 
 const FilterLinks = () => (
   <p>
